Update timeout for automatic edit stop call
This is an issue when we have many items getting customDisplay called on them and webix is trying to identify if you want to edit a cell or stop editing a cell. Increasing the value allows the user to still edit select lists where there are many selectivity renders happening. In the future we will want to just remove customDisplay calls and replace with native webix tools to avoid this.